% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{calcAIC}
\alias{calcAIC}
\title{Akaike Information Criterion}
\usage{
calcAIC(model)
}
\arguments{
\item{model}{the model, i.e. an instance of \code{gmm}}
}
\value{
the AIC
}
\description{
Calculates the AIC with respect to the given \code{model}. In general, the
AIC is defined as \eqn{2 \cdot k - 2 \ln L} where \eqn{k} denotes the
number of free model parameters and \eqn{L} denotes the maximized value of
the likelihood for the given model.
}

