% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{calcAICc}
\alias{calcAICc}
\title{Akaike Information Criterion (corrected)}
\usage{
calcAICc(model, N)
}
\arguments{
\item{model}{the model, i.e. an instance of \code{gmm}}

\item{N}{the number of samples of the dataset used for training}
}
\value{
the AICc
}
\description{
Calculates the AICc with respect to the given \code{model}. In general, the
AICc is defined as \deqn{2 \cdot k - 2 \ln L + \frac{2 \cdot k(k+1)}{N-k-1} =
\textit{AIC} + \frac{2 \cdot k(k+1)}{N-k-1}} where \eqn{k} denotes the number
of free model parameters and \eqn{L} denotes the maximized value of the
likelihood for the given model.
}

