% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{calcBIC}
\alias{calcBIC}
\title{Bayesian Information Criterion}
\usage{
calcBIC(model, N)
}
\arguments{
\item{model}{the model, i.e. an instance of \code{gmm}}

\item{N}{the number of samples of the dataset used for training}
}
\value{
the BIC
}
\description{
Calculates the BIC with respect to the given \code{model}. In general, the
BIC is defined as \eqn{-2 \cdot \ln L + k \cdot \ln N} where \eqn{k} denotes
the number of free model parameters, \eqn{k} denotes the number of free model
parameters, and \eqn{L} denotes the maximized value of the likelihood for the
given model.
}

