% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{computeGMM}
\alias{computeGMM}
\title{EM estimation of a Gaussian Mixture Model.}
\usage{
computeGMM(X, K = 2, lambda = NULL, mu = NULL, Sigma = NULL,
  epsilon = 0.05, maxIter = 250, verb = TRUE)
}
\arguments{
\item{X}{a N x M matrix with N samples and M variables}

\item{K}{an integer specifying the number of Gaussians}

\item{lambda}{a vector of real mixing coefficients}

\item{mu}{a list of mean vectors}

\item{Sigma}{a list of covariance matrices}

\item{epsilon}{the covergence criterion (0 < epsilon < 1), iteration will
stop once last and current likelihood differ by less than \code{epsilon *
100} percent}

\item{maxIter}{the maximum number of iterations}

\item{verb}{be verbose}
}
\value{
an instance of \code{gmm} whose parameters represent the model
 with the best log-likelihood, i.e.
 \item{lambda}{a vector of real mixing coefficients}
 \item{mu}{a list of mean vectors}
 \item{Sigma}{a list of covariance matrices}
 \item{llh}{a vector of the log-likelihood for every iteration step}
}
\description{
EM estimation of a Gaussian Mixture Model that fits the given dataset
\code{X}.
}

