% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{computeSWGMM}
\alias{computeSWGMM}
\title{EM estimation of a (Semi-)Wrapped Gaussian Mixture Model.}
\usage{
computeSWGMM(X, wraps, K = 2, lambda = NULL, mu = NULL, Sigma = NULL,
  trafo = diag(NCOL(X)), invTrafo = NULL, epsilon = 0.05, maxIter = 250,
  verb = TRUE)
}
\arguments{
\item{X}{a N x M matrix with N samples and M variables}

\item{wraps}{an integer vector of length \code{ncol(X)}, describing the
number of wraps per variable, e.g. \code{c(2,0,1)} for a data set with two
periodic variables of which the first should be wrapped twice whereas the
last should be wrapped only once}

\item{K}{an integer specifying the number of Gaussians}

\item{lambda}{a vector of real mixing coefficients}

\item{mu}{a list of mean vectors}

\item{Sigma}{a list of covariance matrices}

\item{trafo}{linear transformation of the displacement grid, defaults to
identity}

\item{invTrafo}{inverse of \code{trafo}. If \code{invTrafo} is \code{NULL}
it will be computed via QR decomposition}

\item{epsilon}{the covergence criterion (0 < epsilon < 1), iteration will
stop once last and current likelihood differ by less than \code{epsilon *
100} percent}

\item{maxIter}{the maximum number of iterations}

\item{verb}{be verbose}
}
\value{
an instance of \code{swgmm} (also: \code{gmm}) whose parameters
represent the model with the best log-likelihood, i.e.
 \item{lambda}{a vector of mixing coefficients}
 \item{mu}{a list of mean vectors}
 \item{Sigma}{a list of covariance matrices}
 \item{wraps}{a copy of the given \code{wraps}}
 \item{wGrid}{for internal use}
 \item{llh}{a vector of the log-likelihood for every iteration step}
}
\description{
EM estimation of a (Semi-)Wrapped Gaussian Mixture Model that fits the given
dataset \code{X}.  Means, covariance matrices and mixing coefficients can be
given e.g. as an initial guess.\cr
Periodic variables will be considered \eqn{2\pi}-periodic. The number of wraps
can be specified per variable through the use of \code{wraps}.\cr
\code{trafo} can be given to apply a linear transformation to the computed
displacement grid, e.g. for whitened datasets whose axes are not aligned with
the original axes anymore.
}
\seealso{
For further details, in particular concerning the displacement grid,
see \code{\link{dswmvnorm}}.
}

