% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{degmm}
\alias{degmm}
\title{Differential entropy of the given (Semi-Wrapped) Gaussian Mixture Model.}
\usage{
degmm(model)
}
\arguments{
\item{model}{an instance of \code{gmm} or \code{swgmm}}
}
\value{
the differential entropy
}
\description{
Differential entropy of the given (Semi-Wrapped) Gaussian Mixture Model.
}

