% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{demvnorm}
\alias{demvnorm}
\title{Differential entropy of a multivariate normal distribution.}
\usage{
demvnorm(Sigma)
}
\arguments{
\item{Sigma}{the covariance matrix (or a vector of the matrix diagonal)}
}
\value{
the differential entropy
}
\description{
Differential entropy of a multivariate normal distribution.
}

