% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{dgmm}
\alias{dgmm}
\title{Probability density of a given ((Semi-)Wrapped) Gaussian Mixture Model.}
\usage{
dgmm(X, model)
}
\arguments{
\item{X}{a vector or a matrix where each row holds one sample and the columns
denote the variables}

\item{model}{the model}
}
\value{
a vector holding the probabilities
}
\description{
Probability density of a given ((Semi-)Wrapped) Gaussian Mixture Model.
}

