% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{dmvnorm}
\alias{dmvnorm}
\title{Probability density of a multivariate Gaussian.}
\usage{
dmvnorm(X, mu = NULL, Sigma = NULL)
}
\arguments{
\item{X}{a vector or a matrix where each row holds one sample and the columns
denote the variables. Note that, when given as a vector, the shape of
\code{mu} and \code{Sigma} determine the shape of \code{X}. Also note that,
\emph{as opposed to \code{mixtools:dmvnorm}}, if neither neither \code{mu}
nor \code{Sigma} are given, \code{X} is assumed to be univariate and the
result is equal to that of \code{dnorm}.}

\item{mu}{the distribution's mean vector}

\item{Sigma}{the distribution's covariance matrix}
}
\value{
a vector holding the probabilities
}
\description{
This is a replacement for \code{mixtools::dmvnorm} that tries to be more
reasonable when \code{X} is given as a vector.
}
\examples{
# Univariate Gaussian with mean 3 and stddev 2
dmvnorm(1, mu=3, Sigma=2)

# Univariate Gaussian with mean 3 and stddev 2 (NOTE that the shape of
# the data, when given as a vector, depends on the shape of mu and/or Sigma)
dmvnorm(1:50, mu=3, Sigma=2)

# Multivariate Gaussian with mean c(3,4) and covariance matrix ((3,-2),(-2,4))
dmvnorm(matrix(1:10, ncol=2), mu=c(3,4), Sigma=matrix(c(3,-2,-2,4), ncol=2))

# In the multivariate case, one or both of mu and Sigma determine the shape
# of the samples when given as a vector.
dmvnorm(1:10, mu=c(3,4), Sigma=matrix(c(3,-2,-2,4), ncol=2))
}

