% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{dswmvnorm}
\alias{dswmvnorm}
\title{Probability density of a (semi-)wrapped multivariate Gaussian.}
\usage{
dswmvnorm(X, mu = NULL, Sigma = NULL, wGrid = NULL)
}
\arguments{
\item{X}{a vector or a matrix where each row holds one sample and the columns
denote the variables. Note that, when given as a vector, the shape of
\code{mu} and \code{Sigma} determine the shape of \code{X}. Also note that,
\emph{as opposed to \code{mixtools:dmvnorm}}, if neither neither \code{mu}
nor \code{Sigma} are given, \code{X} is assumed to be univariate and the
result is equal to that of \code{dnorm}.}

\item{mu}{the distribution's mean vector}

\item{Sigma}{the distribution's covariance matrix}

\item{wGrid}{a matrix whose rows represent the displacements along the
periodic dimensions as vertices of the \eqn{D}-dimensional hypercube where
\eqn{D} denotes the dimension of the data (see Details for further
explanations). If \code{is.null(wGrid)}, a default \eqn{2\pi}-periodic grid
will be chosen for all dimensions}
}
\value{
a vector holding the probabilities
}
\description{
The probability density for a (semi-)wrapped multivariate Gaussian for
periodic data.
}
\details{
The probability density for a (semi-)wrapped multivariate Gaussian for
\eqn{2\pi}-periodic data is defined as

\deqn{ p(x) = \frac{1}{\sqrt{(2\pi)^D \cdot |\Sigma|}} \cdot
\sum_{w \in \mathcal{Z}^D} e^{(x + 2\pi w - \mu)^T
\Sigma^{-1} (x + 2\pi w - \mu)} }

where \eqn{D}, \eqn{\mu} and \eqn{\Sigma} denote the data's dimension, mean
and covariance matrix, and \eqn{\mathcal{Z} = -\infty \ldots
\infty} represents the set of integers.

\eqn{w} represents the displacement (aka the wrapping) along the periodic
dimensions as vertices of the \eqn{D}-dimensional hypercube. For example, if
the first 2 out of 5 dimensions of a semi-wrapped multivariate Gaussian were
\eqn{2\pi}-periodic, then \eqn{w \in \{(2\pi k_1, 2\pi k_2, 0, 0, 0)^T | -1
\le k_1, k_2 \le 1\}}, assuming one wrap in every corresponding direction.

Finally, note that as the computational costs are substantial, depending on
the variance of the data, 1 or 2 wraps are usually considered
sufficient. Also, the number of wraps need not be the same per direction.
}

