% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{entropy}
\alias{entropy}
\title{Discrete entropy of a (Semi-Wrapped) Gaussian Mixture Model.}
\usage{
entropy(X, model)
}
\arguments{
\item{X}{a matrix whose rows represent the multivariate samples}

\item{model}{an instance of \code{gmm} or \code{swgmm}}
}
\value{
the discrete entropy (in nats)
}
\description{
Discrete entropy of a (Semi-Wrapped) Gaussian Mixture Model.
}

