% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastCenter.R
\name{fastRowMinus}
\alias{fastRowMinus}
\title{Subtract a given vector from every row of a given matrix.}
\usage{
fastRowMinus(X, mu)
}
\arguments{
\item{X}{a matrix}

\item{mu}{a vector to be subtracted from every row of \code{X}}
}
\value{
a centered matrix
}
\description{
Subtract a given vector from every row of a given matrix.
}

