% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastCenter.R
\name{fastRowPlus}
\alias{fastRowPlus}
\title{Add a given vector to every row of a given matrix.}
\usage{
fastRowPlus(X, mu)
}
\arguments{
\item{X}{a matrix}

\item{mu}{a vector to be subtracted from every row of \code{X}}
}
\value{
a centered matrix
}
\description{
Add a given vector to every row of a given matrix.
}

