% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{invTransform}
\alias{invTransform}
\title{Inverse transformation of a given Gaussian Mixture Model.}
\usage{
invTransform(model, invH, delta)
}
\arguments{
\item{model}{an instance of \code{gmm}}

\item{invH}{the matrix inverse of the original linear transformation}

\item{delta}{a vector to be added to each of the model's means}
}
\value{
an "inversely transformed" model. Note that the log-likelihood
 history \code{llh} of the resulting model will not be removed, however it
 will be marked as invalid by adding a boolean attribute \code{invalid} with
 value \code{TRUE}.
}
\description{
If the data that were used during the EM-estimation of the given model were
subject to some linear transformation `H`, for example due to whitening of
the dataset prior to the EM-estimation, then consequently the model's
parameters reflect that transformation as well.\cr This function exists
merely for convenience. It applies the given inverse transformation
\code{invH} to all of the model's parameters plus it adds the given
\code{delta} to the model's means.
}

