% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{logdmvnorm}
\alias{logdmvnorm}
\title{Log probability density of a multivariate Gaussian.}
\usage{
logdmvnorm(X, mu = NULL, Sigma = NULL)
}
\arguments{
\item{X}{a vector or a matrix where each row holds one sample and the columns
denote the variables. Note that, when given as a vector, the shape of
\code{mu} and \code{Sigma} determine the shape of \code{X}. Also note that,
\emph{as opposed to \code{mixtools:dmvnorm}}, if neither neither \code{mu}
nor \code{Sigma} are given, \code{X} is assumed to be univariate and the
result is equal to that of \code{dnorm}.}

\item{mu}{the distribution's mean vector}

\item{Sigma}{the distribution's covariance matrix}
}
\value{
a vector holding the logarithm of the probabilities
}
\description{
Log probability density of a multivariate Gaussian.
}
\seealso{
\code{\link{dmvnorm}}
}

