% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-space.R
\name{logSum}
\alias{logSum}
\title{The logarithm of the sum of `a` and `b` where both are given in log-space.}
\usage{
logSum(a, b)
}
\arguments{
\item{a}{a vector of values in log-space}

\item{b}{a vector of values in log-space}
}
\value{
the sum of `a` and `b` in log-space
}
\description{
Proof:
\deqn{
\begin{array}{rl}
\ln(a + b) &= \ln a + \ln(a+b) - \ln a \\
&= \ln a + \ln\frac{a+b}{a} \\
&= \ln a + \ln (1 + \frac{b}{a}) \\
&= \ln a + \ln (1 + e^{\ln b - \ln a})
\end{array}
}
}
\examples{
all.equal( log(7), logSum(log(3), log(4)) )
all.equal( log(7), logSum(log(4), log(3)) )
all.equal( log(7), logSum(log(0), log(7)) )
all.equal( log(7), logSum(log(7), log(0)) )
all.equal( log(0), logSum(log(0), log(0)) )
all.equal( Inf   , logSum(Inf, Inf)       )
all.equal( Inf   , logSum(Inf, -Inf)      )
all.equal( Inf   , logSum(-Inf, Inf)      )
all.equal( Inf   , logSum(Inf, log(3))    )
all.equal( Inf   , logSum(log(3), Inf)    )
}

