% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\name{mapInf}
\alias{mapInf}
\title{Map finite to infinite coordinates and apply function.}
\usage{
mapInf(x, fun, ...)
}
\arguments{
\item{x}{a vector representing the point where to evaluate \code{fun}}

\item{fun}{the function to be applied}

\item{...}{passed on to \code{fun}}
}
\description{
Performs a non-linear change of variables for each of the given coordinates
from \eqn{[-1,+1]} to \eqn{[-Inf, +Inf]}{[-\infty,+\infty]}.
}
\details{
Rule for integration by substitution (univariate):
\deqn{
  \int_{\phi(a)}^{\phi(b)} f(x) dx = \int_a^b f(\phi(t)) \phi'(t) dt
}

And in the multivariate case:
\deqn{
  \int_{\phi(a)}^{\phi(b)} f(x) dx = \int_a^b f(\phi(t)) | J_{\phi}(t) | dt
}

where \eqn{J_{\phi}} denotes the Jacobian of \eqn{\phi} and consequently
\eqn{| J_{\phi}(t) |} the determinant of the Jacobian of \eqn{\phi} evaluated
at \eqn{t}.
}

