% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\name{mapInfPeriodic}
\alias{mapInfPeriodic}
\title{Map finite to infinite respective 2pi-periodic coordinates and apply function.}
\usage{
mapInfPeriodic(x, periodics, fun, ...)
}
\arguments{
\item{x}{a vector representing the point where to evaluate \code{fun}}

\item{periodics}{the indices of the 2pi-periodic variables in \code{x}}

\item{fun}{the function to be applied}

\item{...}{passed on to \code{fun}}
}
\description{
Performs a non-linear change of variables for each of the given
/non-periodic/ coordinates from \eqn{[-1,+1]} to \eqn{[-Inf,
+Inf]}{[-\infty,+\infty]}, while each of the /periodic/ coordinates is
linearly mapped from \eqn{[-1,+1]} to \eqn{[-pi, +pi]}{[-\pi,+\pi]}.
}
\seealso{
\code{\link{mapInf}}
}

