% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{marginal}
\alias{marginal}
\title{Marginal distribution of a (Semi-Wrapped) Gaussian Mixture Model.}
\usage{
marginal(model, margin)
}
\arguments{
\item{model}{an instance of \code{gmm}}

\item{margin}{an integer vector of the indices of the variables to keep}
}
\value{
a marginalized copy of the given \code{model}. Note that the
 log-likelihood history \code{llh} of the resulting model will not be
 removed, however it will be marked as invalid by adding a boolean attribute
 \code{invalid} with value \code{TRUE}.\cr If the given \code{model} is an
 instance of \code{swgmm}, an instance of \code{swgmm} will be returned if
 and only if the marginalized model has any \code{wraps > 0}.
}
\description{
Marginal distribution of the given SW-GMM \code{model}. Only those variables
with corresponding indices in \code{margin} are kept.
}
\examples{
# Create example model.
model <- structure(
  list(mu=list(c(-14, -14), c(0, 0), c(15, -15)),
       Sigma=list(35*diag(2), 25*diag(2), 18*diag(2)),
       lambda=c(0.3, 0.5, 0.2)),
  class="gmm"
)
integrate(function(x) dgmm(x, marginal(model, 1)), -Inf, +Inf)
}

