% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.gmm}
\alias{plot.gmm}
\title{Multidimensional contour plot of a given Gaussian Mixture Model.}
\usage{
\method{plot}{gmm}(x, dataset = NULL, limits = NULL, resolution = 100,
  col = "magenta", oma = c(2, 1, 3.5, 1), ...)
}
\arguments{
\item{x}{the model (an instance of \code{gmm})}

\item{dataset}{a matrix or data-frame whose rows each contain one
d-dimensional observation where d corresponds to the dimensionality of the
given \code{x}. When given as a vector, the vector will be converted to
a one-column matrix. May be \code{NULL} as long as \code{limits} is given.}

\item{limits}{a list of 2-tuples, each describing the extents of the
corresponding axis. May be \code{NULL} as long as \code{dataset} is given. In
the latter case, the axes' extents are derived from the columns of
\code{dataset}.}

\item{resolution}{a vector describing the resolution per axis}

\item{col}{the color to use for the contour plot}

\item{oma}{the outer margin, see \code{par}}

\item{...}{passed on to the \code{contour} function}
}
\description{
All possible combinations of the model's axes are layed out in a grid where
each cell is then filled with a two-dimensional contour plot of the given
GMM. A corresponding two-dimensional histogram will be plotted if a dataset
is provided.
}

