\name{RGMM-package}
\alias{RGMM-package}
%\alias{Gmedian}
\docType{package}
\title{
\packageTitle{RGMM}
}
\description{
In this package, we provide functions to provide robust clustering in the case of Gaussian, Student and Laplace Mixture Models. Function \code{\link{RobVar}} computes robustly the covariance of a numerical data set which are realizations of Gaussian, Student or Laplace vectors. Function \code{\link{RobMM}} enables to provide a clustering of a numerical data set, \code{\link{RMMplot}} enables to produce graph for Robust Mixture Models,  while  \code{\link{Gen_MM}} enables to generate possibly contaminated mixture of Gaussian, Student and Laplace vectors.
%\packageDescription{Gmedian}
}

\author{
\packageAuthor{RGMM}

Maintainer: \packageMaintainer{RGMM}
}
\references{
Cardot, H., Cenac, P. and Zitt, P-A. (2013). Efficient and fast estimation of the geometric median in Hilbert spaces with an averaged stochastic gradient algorithm. \emph{Bernoulli}, 19, 18-43.

Cardot, H. and Godichon-Baggioni, A. (2017). Fast Estimation of the Median Covariation Matrix with Application to Online Robust Principal Components Analysis.  \emph{Test}, 26(3), 461-480

Vardi, Y. and Zhang, C.-H. (2000). The multivariate L1-median and associated data depth. \emph{Proc. Natl. Acad. Sci. USA}, 97(4):1423-1426.

}
\keyword{Robust Mixture Model}
