% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R, R/sampling.R
\docType{package}
\name{rgmm}
\alias{rgmm}
\alias{rgmm-package}
\title{EM estimation and plotting for (Semi-Wrapped) Gaussian Mixture Models}
\usage{
rgmm(N, model)
}
\arguments{
\item{N}{the number of samples to be drawn}

\item{model}{an instance of \code{gmm}}
}
\value{
a matrix whose rows contain the samples drawn from the given
 \code{model}
}
\description{
EM estimation and plotting for (Semi-Wrapped) Gaussian Mixture Models

Draw samples from a given ((Semi-)Wrapped) Gaussian Mixture Model.
}
\author{
Alexander Lehmann <afwlehmann@googlemail.com>
}

