% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Draw random samples from a multivariate Gaussian.}
\usage{
rmvnorm(N, mu = NULL, Sigma = NULL)
}
\arguments{
\item{N}{the number of samples}

\item{mu}{the mean. If \code{mean} is \code{NULL}, zero mean is assumed for
every variable}

\item{Sigma}{the covariance matrix or a vector of variances. If \code{Sigma}
is \code{NULL}, unit variance is assumed for every variable}
}
\value{
a matrix with one sample per row
}
\description{
Draw random samples from a multivariate Gaussian.
}

